Final file/directory rearrangement.
- Still needs linting!
