+++ src/pyibisami/common.py
@@ -19,6 +19,7 @@ Cvec: TypeAlias = npt.NDArray["Comp"]
 
 PI:    float = 3.141592653589793238462643383279502884
 TWOPI: float = 2.0 * PI
+EPS:   float = 0.0001  # Used to test floats for "== 0".
 
 # TestConfig: TypeAlias = tuple[str, tuple[dict[str, Any], dict[str, Any]]]
 # TestSweep:  TypeAlias = tuple[str, str, list[TestConfig]]
