+++ src/pyibisami/testing/ibis_file_tests.py
@@ -17,9 +17,8 @@ from reportlab.platypus import Flowable, Image, ListFlowable, ListItem, PageBrea
 
 from ..common import TestSweep
 from ..ibis.file import IBISModel
-from ..util.reportlab_combinators import preformatted
-from ..util.tool_helpers import (
-    init_vs_getwave, plot_sweeps, samples_per_bit, check_getwave_input_length,
+from ..util.reportlab import (
+    preformatted,
     bold, ital, fixed, page_break, spacer,
     styles, bold_style, caption_style, indented_style,
     P, H1, H2, H3, H4)
