+++ src/pyibisami/tools/__init__.py
@@ -1,0 +1,5 @@
+"""
+WARNING: THIS MODULE IS DEPRECATED!
+
+Please, use the new testing infrastructure in ``pyibisami.testing`` instead.
+"""
